Use error-prone compiler; fix bugs and resource leaks
