+++ pom.xml
@@ -282,21 +282,36 @@
         <artifactId>maven-compiler-plugin</artifactId>
         <version>3.7.0</version>
         <configuration>
+          <compilerId>javac-with-errorprone</compilerId>
+          <forceJavacCompilerUse>true</forceJavacCompilerUse>
           <source>${java.src.version}</source>
           <target>${java.src.version}</target>
           <testSource>${java.test.version}</testSource>
           <testTarget>${java.test.version}</testTarget>
           <showWarnings>true</showWarnings>
           <compilerArgs>
-              <arg>-Xlint:deprecation,unchecked</arg>
-              <!-- always generate class files for package-info -->
-              <arg>-Xpkginfo:always</arg>
+            <arg>-Xlint:deprecation,unchecked</arg>
+            <!-- always generate class files for package-info -->
+            <arg>-Xpkginfo:always</arg>
+            <arg>-XepExcludedPaths:(.*/testclasses.*/.*|.*/Foo.*)</arg>
           </compilerArgs>
           <!-- incremental compilation is broken in maven, see https://issues.apache.org/jira/browse/MCOMPILER-205  -->
           <!-- useIncrementalCompilation=false actually means that we won't recompile everything each time -->
           <!-- it's not a joka, test by yourself :-) -->
           <useIncrementalCompilation>false</useIncrementalCompilation>
         </configuration>
+        <dependencies>
+          <dependency>
+            <groupId>org.codehaus.plexus</groupId>
+            <artifactId>plexus-compiler-javac-errorprone</artifactId>
+            <version>2.8.3</version>
+          </dependency>
+          <dependency>
+            <groupId>com.google.errorprone</groupId>
+            <artifactId>error_prone_core</artifactId>
+            <version>2.2.0</version>
+          </dependency>
+        </dependencies>
       </plugin>
 
       <plugin>
