+++ src/main/java/spoon/support/StandardEnvironment.java
@@ -47,6 +47,7 @@ import java.net.URL;
 import java.net.URLClassLoader;
 import java.nio.charset.Charset;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
@@ -398,7 +399,7 @@ public class StandardEnvironment implements Serializable, Environment {
 			try {
 				urls[i] = new File(classpath[i]).toURI().toURL();
 			} catch (MalformedURLException e) {
-				throw new IllegalStateException("Invalid classpath: " + classpath, e);
+				throw new IllegalStateException("Invalid classpath: " + Arrays.toString(classpath), e);
 			}
 		}
 		return urls;
