+++ src/test/java/spoon/test/SpoonTestHelpers.java
@@ -1,11 +1,8 @@
 package spoon.test;
 
-import spoon.Launcher;
-import spoon.SpoonAPI;
 import spoon.reflect.declaration.CtElement;
 import spoon.reflect.declaration.CtMethod;
 import spoon.reflect.declaration.CtType;
-import spoon.reflect.declaration.ModifierKind;
 import spoon.reflect.reference.CtTypeReference;
 import spoon.reflect.visitor.filter.OverridingMethodFilter;
 import spoon.support.DerivedProperty;
@@ -87,7 +84,7 @@ public class SpoonTestHelpers {
 				//parent is a special kind of setter, which does not influence model properties of element, but link to parent element.
 				continue;
 			}
-			if (!m.getSimpleName().startsWith("set") && !m.getSimpleName().startsWith("set")) {
+			if (!m.getSimpleName().startsWith("set")) {
 				continue;
 			}
 			if (m.getParameters().size()!=1) {
