+++ src/test/java/spoon/test/control/testclasses/DoWhile.java
@@ -1,4 +1,4 @@
-package spoon.test.control;
+package spoon.test.control.testclasses;
 
 public class DoWhile {
