+++ src/test/java/spoon/test/processing/testclasses/SampleForInsertBefore.java
@@ -1,4 +1,4 @@
-package spoon.test.processing;
+package spoon.test.processing.testclasses;
 
 public class SampleForInsertBefore {
