+++ src/test/java/spoon/test/properties/PropertiesTest.java
@@ -24,7 +24,7 @@ public class PropertiesTest {
 				factory,
 				SpoonResourceHelper
 						.resources(
-								"./src/test/java/spoon/test/properties/Sample.java"
+								"./src/test/java/spoon/test/properties/testclasses/Sample.java"
 						));
 		compiler.build();
